crate.settings: Extract `crate.settings.index` route
