+++ app/routes/crate/settings.js
@@ -5,7 +5,6 @@ import AuthenticatedRoute from '../-authenticated-route';
 export default class SettingsRoute extends AuthenticatedRoute {
   @service router;
   @service session;
-  @service store;
 
   async beforeModel(transition) {
     await super.beforeModel(...arguments);
@@ -19,20 +18,5 @@ export default class SettingsRoute extends AuthenticatedRoute {
         title: 'This page is only accessible by crate owners',
       });
     }
-  }
-
-  async model() {
-    let crate = this.modelFor('crate');
-
-    let githubConfigs = await this.store.query('trustpub-github-config', { crate: crate.name });
-
-    return { crate, githubConfigs };
-  }
-
-  setupController(controller, { crate, githubConfigs }) {
-    super.setupController(...arguments);
-
-    controller.set('crate', crate);
-    controller.set('githubConfigs', githubConfigs);
   }
 }
